Lights applied
Light does not work in a 3D world, but it will work in a 3D project.

* Created a new diffuse material for sprites and applied it to all
prefabs that needs it.
* Added a spotlight and a point light to the player. The spotlight is
the flashlight and the point light is the spot on the player.
As long as the player is rotated the light will rotate with the player.
Does not yet work for orientation, but the flash light can be oriented
via its rotation.